sm6225-common: Remove software c2 codec overrides
For some reason, we fail to generate thumbnails after 15 QPR1 using the
hardware OMX decoders. Remove the limits of the software codecs to make
the thumbnails work again, although this might cause performance issues
and possibly broken tests. Until we have a better solution, we have to
live with this.

Change-Id: I6ce3a99db72753972d4f1323371dac9b5f08bf6b